teacher_info_search_state.py renamed into teacher_info_search_states.py
